\name{TrecFastTest-package}
\alias{TrecFastTest-package}
\alias{TrecFastTest}
\docType{package}
\title{
\packageTitle{TrecFastTest}
}
\description{
\packageDescription{TrecFastTest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TrecFastTest}
\packageIndices{TrecFastTest}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TrecFastTest}

Maintainer: \packageMaintainer{TrecFastTest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
